+++ pkg/plugin/chrome/stream.go
