+++ pkg/plugin/chrome/tab.go
