+++ pkg/plugin/client/errors.go
