+++ pkg/plugin/config/settings.go
