+++ pkg/plugin/config/settings_test.go
