+++ pkg/plugin/dashboard/errors.go
