+++ pkg/plugin/report/templates/footer.gohtml
