+++ pkg/plugin/report/templates/header.gohtml
