+++ pkg/plugin/worker/main.go
